requirements.txt
playwright
pytest
pytest-html
pytest-xdist
openpyxl
allure-pytest